+++ src/operations/event_generation_helpers.cpp
@@ -1,0 +1,123 @@
+// Copyright (c) 2023-2024 Percona and/or its affiliates.
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License, version 2.0,
+// as published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License, version 2.0, for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
+
+#include "operations/event_generation_helpers.hpp"
+
+#include <cstdint>
+
+#include "binsrv/events/checksum_algorithm_type.hpp"
+#include "binsrv/events/code_type.hpp"
+#include "binsrv/events/common_header_flag_type.hpp"
+#include "binsrv/events/composite_binlog_name_fwd.hpp"
+#include "binsrv/events/event.hpp"
+#include "binsrv/events/event_view.hpp"
+#include "binsrv/events/protocol_traits_fwd.hpp"
+#include "binsrv/events/reader_context.hpp"
+
+#include "binsrv/gtids/gtid_set_fwd.hpp"
+
+#include "util/ctime_timestamp.hpp"
+#include "util/semantic_version.hpp"
+
+namespace operations {
+
+[[nodiscard]] binsrv::events::event_view generate_rotate_event(
+    binsrv::events::event_storage &event_buffer,
+    const binsrv::events::reader_context &context, std::uint32_t offset,
+    bool current_timestamp, std::uint32_t server_id, bool artificial,
+    const binsrv::events::composite_binlog_name &binlog_name) {
+  const binsrv::events::generic_post_header<binsrv::events::code_type::rotate>
+      post_header{binsrv::events::magic_binlog_offset};
+  const binsrv::events::generic_body<binsrv::events::code_type::rotate> body{
+      binlog_name};
+
+  util::ctime_timestamp timestamp{};
+  if (current_timestamp) {
+    timestamp = util::ctime_timestamp::now();
+  }
+
+  binsrv::events::common_header_flag_set flags{};
+  if (artificial) {
+    flags |= binsrv::events::common_header_flag_type::artificial;
+  }
+
+  // the value of the 'include_checksum' parameters is taken from the
+  // 'reader_context': immediately after reconnection it will be equal to
+  // the '<replication.verify_checksum>' configuration parameter and after
+  // that will be taken from the FORMAT_DESCRIPTION events, which in the
+  // rewrite mode will be generated by us and therefore will always include
+  // 'checksum_algorithm' set to 'crc32'
+  const auto generated_event{
+      binsrv::events::event::create_event<binsrv::events::code_type::rotate>(
+          offset, timestamp, server_id, flags, post_header, body,
+          context.is_footer_expected(), event_buffer)};
+
+  return binsrv::events::event_view{context,
+                                    util::const_byte_span{event_buffer}};
+}
+
+[[nodiscard]] binsrv::events::event_view
+generate_format_description_event(binsrv::events::event_storage &event_buffer,
+                                  const binsrv::events::reader_context &context,
+                                  std::uint32_t offset,
+                                  std::uint32_t server_id) {
+  const util::semantic_version server_version{
+      context.get_current_encoded_server_version()};
+  const binsrv::events::generic_post_header<
+      binsrv::events::code_type::format_description>
+      post_header{
+          binsrv::events::default_binlog_version, server_version,
+          util::ctime_timestamp::now(),
+          binsrv::events::default_common_header_length,
+          binsrv::events::reader_context::get_hardcoded_post_header_lengths(
+              server_version.get_encoded())};
+  const binsrv::events::generic_body<
+      binsrv::events::code_type::format_description>
+      body{binsrv::events::checksum_algorithm_type::crc32};
+
+  // enforcing checksums for all rewritten upcoming events
+  const auto generated_event{binsrv::events::event::create_event<
+      binsrv::events::code_type::format_description>(
+      offset, util::ctime_timestamp::now(), server_id,
+      binsrv::events::common_header_flag_set{}, post_header, body,
+      true /* include_checksum */, event_buffer)};
+
+  return binsrv::events::event_view{context,
+                                    util::const_byte_span{event_buffer}};
+}
+
+[[nodiscard]] binsrv::events::event_view
+generate_previous_gtids_log_event(binsrv::events::event_storage &event_buffer,
+                                  const binsrv::events::reader_context &context,
+                                  std::uint32_t offset, std::uint32_t server_id,
+                                  const binsrv::gtids::gtid_set &gtids) {
+  const binsrv::events::generic_post_header<
+      binsrv::events::code_type::previous_gtids_log>
+      post_header{};
+  const binsrv::events::generic_body<
+      binsrv::events::code_type::previous_gtids_log>
+      body{gtids};
+  const auto generated_previous_gtids_log_event{
+      binsrv::events::event::create_event<
+          binsrv::events::code_type::previous_gtids_log>(
+          offset, util::ctime_timestamp::now(), server_id,
+          binsrv::events::common_header_flag_set{}, post_header, body, true,
+          event_buffer)};
+
+  return binsrv::events::event_view{context,
+                                    util::const_byte_span{event_buffer}};
+}
+
+} // namespace operations
